Move api into a more robust structure, make parsing 'less complex'
